pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deploy-guard"
version = "0.1.0"
description = "Autonomous Deployment Validator (AI + CI/CD)"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [{name = "Deploy Guard Authors"}]
dependencies = [
  "pyyaml>=6.0.1",
  "click>=8.1.7",
  "rich>=13.7.1",
  "python-hcl2>=5.1.1",
]

[project.optional-dependencies]
llm = ["openai>=1.40.0", "requests>=2.32.0"]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.6.0",
  "mypy>=1.11.1",
]

[project.scripts]
deploy-guard = "deploy_guard.cli:main"

[tool.setuptools]
packages = ["deploy_guard", "deploy_guard.scanners", "deploy_guard.rules", "deploy_guard.llm", "deploy_guard.reporting"]

[tool.ruff]
line-length = 260
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["UP017"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
exclude = ["demo/"]

[[tool.mypy.overrides]]
module = ["hcl2", "openai", "requests"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["deploy_guard.llm.providers"]
ignore_errors = true
